Fix SCB installment charge test
